end of the course

+++ resources/item.py
@@ -1,5 +1,11 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+                                jwt_required,
+                                get_jwt_claims,
+                                jwt_optional,
+                                get_jwt_identity,
+                                fresh_jwt_required
+                                )
 from models.item import ItemModel
 
 class Item(Resource):
@@ -18,14 +24,14 @@ class Item(Resource):
         help="Every item needs a store id"
     )
 
-    @jwt_required()
+    @jwt_required
     def get(self,name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {'message': 'Item Not Found'},404
 
-
+    @fresh_jwt_required
     def post(self,name):
         if ItemModel.find_by_name(name):
             return {'message': f"an item with {name} already exists."}, 400
@@ -40,7 +46,11 @@ class Item(Resource):
 
         return item.json(),201
 
+    @jwt_required
     def delete(self,name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message':'admin privilege required'}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -60,5 +70,27 @@ class Item(Resource):
         return item.json()
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [item.json() for item in ItemModel.find_all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items}, 200
+        return {'items': [item['name'] for item in items],
+                'message': 'more data available if you log in.'
+        }, 200
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+#d
